+++ software/doom/src/sdcard/SD.h
@@ -41,7 +41,7 @@ namespace SDLib {
       virtual int peek();
       virtual int available();
       virtual void flush();
-      int read(void *buf, uint16_t nbyte);
+      int read(void *buf, size_t nbyte);
       bool seek(uint32_t pos);
       uint32_t position();
       uint32_t size();
